F-35: extract ObservationRelay<Element> from PreferencesObserver and ScopesObserver
PreferencesObserver and ScopesObserver in RunnerPollerObservers.swift were
structurally identical: same continuation + store init pattern, same recursive
withObservationTracking loop. Only the element type and the read/transform
closure differed.

- Add ObservationRelay<Element: Sendable> (new file) — generic @MainActor relay
  that accepts a continuation and a @MainActor () -> Element read closure.
- Replace both concrete classes in RunnerPollerObservers.swift with typealiases
  so call sites in RunnerPoller.swift require no renaming.
- Update the two MainActor.run init blocks in RunnerPoller.swift to use the
  trailing-closure init instead of the old store: parameter.

No behaviour change. Resolves #1648.
